requirements.txt
click==7.1.1
mongoengine==0.19.1
pymongo==3.10.1
PyYAML==5.3.1
six==1.14.0
